Implement & link famicom.cpp

+++ include/m_card.h
@@ -16,7 +16,7 @@ typedef struct {
   PersonalID_c pid[PLAYER_NUM];
 } mCD_persistent_data_c;
 
-extern u8 mCD_GetThisLandSlotNo_code(int* player_no, int* slot_card_results);
+extern s32 mCD_GetThisLandSlotNo_code(int* player_no, s32* slot_card_results);
 extern void mCD_save_data_aram_malloc();
 extern void mCD_set_aram_save_data();
 extern void mCD_init_card();
